+++ dating-plan-ai-agents/environment.yml
@@ -185,9 +185,6 @@ dependencies:
       - threadpoolctl==3.5.0
       - tokenizers==0.20.4
       - toml==0.10.2
-      - torch==2.5.1
-      - tqdm==4.67.1
-      - transformers==4.46.3
       - types-python-dateutil==2.9.0.20241003
       - urllib3==2.2.3
       - uvicorn==0.32.1
